Intial code (#1)
* Inital commit

The code is not clean, I shall clean in upcoming commits. However this way you can interact with the code and get a feel for the layout on your device

* Inital Commit

The code is not cleaned yet, I still need to go the relevant commenting but here it is. This way you can run the demo on your device

* Inital Commit

The code is not cleaned yet, I still need to go the relevant commenting but here it is. This way you can run the demo on your device